Conflicts:
	src/Arch/Arm/ArmRewriter.Alu.cs
	src/Arch/Arm/ArmRewriter.cs
	subjects/Elf-MIPS/redir.c
	subjects/Elf-MIPS/redir.h
	subjects/PDP-11/lunar.c
	subjects/PDP-11/lunar.h
	subjects/PDP-11/space.c
	subjects/PE-x86/ExeDll/Executable.c
	subjects/PE/x86/import-ordinals/main.c

+++ src/Decompiler/Analysis/Aliases.cs
@@ -243,7 +243,7 @@ namespace Reko.Analysis
 				}
 				else
 				{
-					aliasExpr = new Slice(varTo.DataType, varFrom, (uint) offsetFrom);
+					aliasExpr = new Slice(varTo.DataType, varFrom, offsetFrom);
 				}
 			}
 			else
